Rename types module to prevent name conflict.
This change renames the planetlab/types.py module to model.py to prevent
a name conflict for unit tests on the objects in planetlab/types.py (now
model.py).
