+++ plsync/sites.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 
 import pprint
-from planetlab.types import *
+from planetlab.model import *
 from users import user_list
 
 # NOTE: The legacy network remap is used to re-order the automatically
